update : 자판기Archiver 파일로 분리 : 기존 AppDelegate에 있던것 이동, VendingMachine에 private 변수 추가
