+++ VendingMachineApp/VendingMachineApp/AppDelegate.swift
@@ -10,16 +10,9 @@ import UIKit
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
-    var vendingMachine: VendingMachine?
     var window: UIWindow?
 
-
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-        do {
-            try vendingMachine = AppDelegate.load()
-        } catch {
-            vendingMachine = AppDelegate.set()
-        }
 
         return true
     }
@@ -27,7 +20,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func applicationWillResignActive(_ application: UIApplication) { }
 
     func applicationDidEnterBackground(_ application: UIApplication) {
-        AppDelegate.archive(vendingMachine: vendingMachine)
+        VMArchiver.archive()
     }
 
     func applicationWillEnterForeground(_ application: UIApplication) { }
@@ -35,32 +28,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func applicationDidBecomeActive(_ application: UIApplication) { }
 
     func applicationWillTerminate(_ application: UIApplication) {
-        AppDelegate.archive(vendingMachine: vendingMachine)
-    }
-    
-    static func archive(vendingMachine: VendingMachine?) {
-        guard let vendingMachine = vendingMachine else { return }
-        let vendingMachineEncoded = try? NSKeyedArchiver.archivedData(
-            withRootObject: vendingMachine,
-            requiringSecureCoding: false)
-        UserDefaults.standard.set(vendingMachineEncoded, forKey:"vendingMachine")
-        
-    }
-    
-    enum loadError: Error {
-        case noData
-        case noLoad
-    }
-    
-    static func load() throws -> VendingMachine? {
-        guard let data = UserDefaults.standard.data(forKey: "vendingMachine") else { throw loadError.noData }
-        guard let vendingMachine = try NSKeyedUnarchiver
-            .unarchiveTopLevelObjectWithData(data) as? VendingMachine else { throw loadError.noLoad }
-        return vendingMachine
-    }
-
-    static func set() -> VendingMachine {
-        return VendingMachine.init(list: Inventory(list: [ObjectIdentifier: Packages]()))
+        VMArchiver.archive()
     }
     
 }
